Move RPC codec tests into serialization suite
